@@ -1,27 +1,22 @@
 {
   "name": "gofar",
-  "version": "2.0.0",
+  "version": "3.0.0",
   "description": "GO-FAR score calculator.",
   "author": "Jonathan Rubins <jonrubins@gmail.com>",
   "scripts": {
     "build": "NODE_ENV=production webpack --progress --config webpack.config.prod.js",
-    "bundle:analyze": "NODE_ENV=production webpack --progress --config webpack.config.prod.js --profile --json > stats.json && webpack-bundle-analyzer stats.json",
-    "bundle:analyze:dev": "webpack --progress --config webpack.config.dev.js --profile --json > stats.json && webpack-bundle-analyzer stats.json",
     "lint:js": "eslint app buildConfig.js postcss.config.js webpack.config.dev.js webpack.config.prod.js --cache --ext=.js,.jsx",
     "lint:sass": "sass-lint -v -q",
     "lint": "yarn lint:js && yarn lint:sass",
-    "precommit": "yarn lint && yarn test",
     "start": "git pull && yarn start:merge",
     "start:merge": "yarn && yarn start:offline",
     "start:debug": "DEBUG=true && yarn start",
-    "start:offline": "webpack-dev-server --progress --config webpack.config.dev.js",
-    "test": "jest --passWithNoTests",
-    "test:watch": "jest --watch"
+    "start:offline": "webpack-dev-server --progress --config webpack.config.dev.js"
   },
   "engines": {
-    "node": "8.6.0",
-    "npm": "5.3.0",
-    "yarn": "1.3.2"
+    "node": "8.11.4",
+    "npm": "5.6.0",
+    "yarn": "1.9.4"
   },
   "repository": {
     "type": "git",
@@ -29,54 +24,44 @@
   },
   "license": "MIT",
   "private": true,
-  "jest": {
-    "cacheDirectory": "./.tmp",
-    "resetMocks": true,
-    "resetModules": true
-  },
   "dependencies": {
-    "classnames": "2.2.5",
+    "@jrubins/react-components": "0.0.7",
+    "@jrubins/utils": "0.0.6",
+    "classnames": "2.2.6",
     "isomorphic-fetch": "2.2.1",
-    "lodash": "4.17.5",
-    "prop-types": "15.6.0",
-    "react": "16.2.0",
-    "react-dom": "16.2.0",
-    "react-redux": "5.0.7",
-    "react-router-dom": "4.2.2",
-    "redux": "3.7.2",
-    "redux-thunk": "2.2.0"
+    "lodash": "4.17.11",
+    "prop-types": "15.6.2",
+    "react": "16.7.0-alpha.2",
+    "react-dom": "16.7.0-alpha.2"
   },
   "devDependencies": {
-    "autoprefixer": "7.2.4",
-    "babel-core": "6.26.0",
-    "babel-eslint": "8.2.2",
-    "babel-loader": "7.1.3",
-    "babel-plugin-lodash": "3.3.2",
-    "babel-plugin-transform-es2015-destructuring": "6.23.0",
-    "babel-plugin-transform-object-rest-spread": "6.26.0",
-    "babel-polyfill": "6.26.0",
-    "babel-preset-env": "1.6.1",
-    "babel-preset-react": "6.24.1",
-    "copy-webpack-plugin": "4.4.2",
-    "css-loader": "0.28.10",
-    "dotenv": "5.0.0",
-    "eslint": "4.18.1",
-    "eslint-plugin-react": "7.7.0",
-    "extract-text-webpack-plugin": "3.0.2",
-    "html-webpack-plugin": "2.30.1",
-    "husky": "0.14.3",
-    "jest": "22.4.2",
-    "nock": "9.2.1",
-    "node-sass": "4.7.2",
-    "postcss-loader": "2.1.0",
-    "react-hot-loader": "4.0.0-beta.14",
-    "redux-mock-store": "1.5.1",
+    "@babel/core": "7.1.6",
+    "@babel/polyfill": "7.0.0",
+    "@babel/preset-env": "7.1.6",
+    "@babel/preset-react": "7.0.0",
+    "autoprefixer": "9.3.1",
+    "babel-eslint": "10.0.1",
+    "babel-loader": "8.0.4",
+    "babel-plugin-lodash": "3.3.4",
+    "copy-webpack-plugin": "4.6.0",
+    "css-loader": "1.0.1",
+    "dotenv": "6.1.0",
+    "eslint": "5.9.0",
+    "eslint-plugin-react": "7.11.1",
+    "html-webpack-plugin": "3.2.0",
+    "husky": "1.2.0",
+    "lint-staged": "8.1.0",
+    "mini-css-extract-plugin": "0.4.5",
+    "node-sass": "4.10.0",
+    "postcss-loader": "3.0.0",
+    "prettier": "1.15.2",
+    "react-hot-loader": "4.3.12",
     "sass-lint": "1.12.1",
-    "sass-loader": "6.0.6",
-    "style-loader": "0.20.2",
-    "webpack": "3.10.0",
-    "webpack-bundle-analyzer": "2.11.0",
-    "webpack-dev-server": "2.9.7",
-    "webpack-dotenv-plugin": "2.0.2"
+    "sass-loader": "7.1.0",
+    "style-loader": "0.23.1",
+    "webpack": "4.26.0",
+    "webpack-cli": "3.1.2",
+    "webpack-dev-server": "3.1.10",
+    "webpack-dotenv-plugin": "2.1.0"
   }
 }
