@@ -1,13 +1,10 @@
-import { Component } from 'react'
+import { useEffect } from 'react'
+import { insertScript } from '@jrubins/utils/lib/dom'
 
-import {
-  ANALYTICS_LIBS,
-  setupAnalytics,
-} from '../../../utils/analytics'
-import { insertScript } from '../../../utils/dom'
+import { ANALYTICS_LIBS, setupAnalytics } from '../../../utils/analytics'
 
-class Ga extends Component {
-  componentDidMount() {
+const Ga = () => {
+  useEffect(() => {
     // Setup our analytics data BEFORE we insert the actual script onto the DOM.
     setupAnalytics(ANALYTICS_LIBS.GA)
 
@@ -15,15 +12,9 @@ class Ga extends Component {
       id: 'ga-sdk',
       src: 'https://www.google-analytics.com/analytics.js',
     })
-  }
+  })
 
-  shouldComponentUpdate() {
-    return false
-  }
-
-  render() {
-    return null
-  }
+  return null
 }
 
 export default Ga
