@@ -1,66 +1,76 @@
-import React from 'react'
+import React, { useState } from 'react'
 import PropTypes from 'prop-types'
-import { connect } from 'react-redux'
+import _ from 'lodash'
 
-import { MODAL_TYPES } from '../../../utils/modals'
+import { SYMPTOMS } from '../../../utils/symptoms'
 import { customEvent } from '../../../utils/analytics'
-
-import { openModal } from '../../../actions/modal'
 
 import Attribution from '../../reusable/attribution/Attribution'
 import PatientAgeForm from '../../reusable/patients/PatientAgeForm'
 import ScoreOutput from '../../reusable/output/ScoreOutput'
 import SymptomsForm from '../../reusable/symptoms/SymptomsForm'
 
-/**
- * Handles when a user clicks the feedback link.
- *
- * @param {Function} openModal
- */
-function handleFeedbackClick(openModal) {
-  customEvent('Feedback', 'Opened', 'Modal')
+const HomePage = ({ openFeedbackModal }) => {
+  const [age, setAge] = useState(null)
+  const [selectedSymptomIds, setSelectedSymptomIds] = useState([])
 
-  openModal({
-    type: MODAL_TYPES.FEEDBACK,
-  })
-}
+  return (
+    <div className="home-page page">
+      <h1 className="gofar-title">
+        Good Outcome Following Attempted Resuscitation (GO-FAR)
+      </h1>
 
-const HomePage = ({ openModal }) => (
-  <div className="home-page page">
-    <h1 className="gofar-title">
-      Good Outcome Following Attempted Resuscitation (GO-FAR)
-    </h1>
+      <p className="check-condition-instructions">
+        Check each condition present on admission to the hospital to calculate
+        total score and probability of survival.
+      </p>
 
-    <p className="check-condition-instructions">
-      Check each condition present on admission to the hospital to calculate total
-      score and probability of survival.
-    </p>
+      <PatientAgeForm age={age} onAgeChanged={setAge} />
 
-    <PatientAgeForm />
+      <SymptomsForm
+        selectedSymptomIds={selectedSymptomIds}
+        toggleSymptom={symptomId => {
+          if (_.includes(selectedSymptomIds, symptomId)) {
+            setSelectedSymptomIds(_.without(selectedSymptomIds, symptomId))
+          } else {
+            setSelectedSymptomIds([...selectedSymptomIds, symptomId])
+          }
+        }}
+      />
 
-    <SymptomsForm />
+      <ScoreOutput
+        age={age}
+        resetCalculator={() => {
+          setAge('')
+          setSelectedSymptomIds([])
+        }}
+        selectedSymptoms={_.filter(SYMPTOMS, ({ id }) =>
+          _.includes(selectedSymptomIds, id)
+        )}
+      />
 
-    <ScoreOutput />
+      <Attribution />
 
-    <Attribution />
+      <div className="footer">
+        <span>Jonathan Rubins &copy; 2014-2018</span>
+        <span className="footer-divider hidden-sm hidden-xs" />
+        <a
+          className="hidden-sm hidden-xs"
+          onClick={() => {
+            customEvent('Feedback', 'Opened', 'Modal')
 
-    <div className="footer">
-      <span>Jonathan Rubins &copy; 2014-2018</span>
-      <span className="footer-divider hidden-sm hidden-xs"></span>
-      <a
-        className="hidden-sm hidden-xs"
-        onClick={() => handleFeedbackClick(openModal)}
-      >
-        Feedback
-      </a>
+            openFeedbackModal()
+          }}
+        >
+          Feedback
+        </a>
+      </div>
     </div>
-  </div>
-)
+  )
+}
 
 HomePage.propTypes = {
-  openModal: PropTypes.func.isRequired,
+  openFeedbackModal: PropTypes.func.isRequired,
 }
 
-export default connect(null, {
-  openModal,
-})(HomePage)
+export default HomePage
