@@ -1,10 +1,10 @@
-import React from 'react'
+import React, { useEffect, useRef, useState } from 'react'
 import PropTypes from 'prop-types'
+import { offset } from '@jrubins/utils/lib/dom'
 import _ from 'lodash'
 import cn from 'classnames'
 
 import { customEvent } from '../../../utils/analytics'
-import { offset } from '../../../utils/dom'
 
 import InfoIcon from '../icons/InfoIcon'
 
@@ -15,63 +15,40 @@ import InfoIcon from '../icons/InfoIcon'
  */
 const INFO_TEXT_WIDTH = 250
 
-class SymptomInfo extends React.Component {
-  constructor(props) {
-    super(props)
-
-    this.state = {
-      arrowLeft: '50%',
-      containerLeft: 0,
-      showInfoText: false,
-    }
-
-    this.mouseOverInfoText = false
-
-    this.handleInfoTextClick = this.handleInfoTextClick.bind(this)
-    this.handleInfoTextMouseEnter = this.handleInfoTextMouseEnter.bind(this)
-    this.handleInfoTextMouseLeave = this.handleInfoTextMouseLeave.bind(this)
-    this.handleWindowResize = this.handleWindowResize.bind(this)
-    this.throttledCalculateLeftShift = _.throttle(this.calculateLeftShift, 300)
-  }
-
-  componentDidMount() {
-    this.throttledCalculateLeftShift()
-
-    window.addEventListener('resize', this.handleWindowResize)
-  }
-
-  componentWillUmount() {
-    window.clearTimeout(this.showInfoTextTimeout)
-
-    window.removeEventListener('resize', this.handleWindowResize)
-  }
+const SymptomInfo = ({ infoText }) => {
+  const symptomInfoContainer = useRef(null)
+  const [arrowLeft, setArrowLeft] = useState('50%')
+  const [containerLeft, setContainerLeft] = useState(0)
+  const [showInfoText, setShowInfoText] = useState(false)
+  const throttledCalculateLeftShift = _.throttle(calculateLeftShift, 300)
+  let mouseOverInfoText = false
+  let showInfoTextTimeout = null
 
   /**
    * Calculates the left shift for the info text box and its arrow indicator taking
    * into account the width of the user's viewport.
    */
-  calculateLeftShift() {
-    const containerOffset = offset(this.symptomInfoContainer)
-    const halfInfoTextWidth = (INFO_TEXT_WIDTH) / 2
+  function calculateLeftShift() {
+    const containerOffset = offset(symptomInfoContainer.current)
+    const halfInfoTextWidth = INFO_TEXT_WIDTH / 2
 
     // Window bounds are adjusted so the info text boxes don't ride up against the side of the screen.
     const windowRightBound = window.innerWidth - 15
     const windowLeftBound = 15
 
-    let leftShift = -1 * (halfInfoTextWidth)
+    let leftShift = -1 * halfInfoTextWidth
     if (containerOffset.left + halfInfoTextWidth > windowRightBound) {
       // Subtract the overlap of the info text box from where the box was going to be positioned.
-      const windowOverlap = containerOffset.left + halfInfoTextWidth - windowRightBound
+      const windowOverlap =
+        containerOffset.left + halfInfoTextWidth - windowRightBound
       leftShift = -1 * (halfInfoTextWidth + windowOverlap)
     } else if (containerOffset.left - halfInfoTextWidth < windowLeftBound) {
       // Position the info text box so its "left" coordinate is 15.
       leftShift = -1 * (containerOffset.left - 15)
     }
 
-    this.setState({
-      arrowLeft: leftShift * -1,
-      containerLeft: leftShift,
-    })
+    setArrowLeft(leftShift * -1)
+    setContainerLeft(leftShift)
   }
 
   /**
@@ -79,13 +56,11 @@ class SymptomInfo extends React.Component {
    *
    * @param {Event} event
    */
-  handleInfoTextClick(event) {
+  function handleInfoTextClick(event) {
     event.preventDefault()
     event.stopPropagation()
 
-    this.setState({
-      showInfoText: !this.state.showInfoText,
-    })
+    setShowInfoText(!showInfoText)
   }
 
   /**
@@ -93,17 +68,15 @@ class SymptomInfo extends React.Component {
    *
    * @param {Event} event
    */
-  handleInfoTextMouseEnter() {
-    this.mouseOverInfoText = true
+  function handleInfoTextMouseEnter() {
+    mouseOverInfoText = true
 
     // Delay showing the info text box so it's not annoying.
-    this.showInfoTextTimeout = window.setTimeout(() => {
-      if (this.mouseOverInfoText) {
+    showInfoTextTimeout = window.setTimeout(() => {
+      if (mouseOverInfoText) {
         customEvent('Info Text', 'Shown')
 
-        this.setState({
-          showInfoText: true,
-        })
+        setShowInfoText(true)
       }
     }, 350)
   }
@@ -113,61 +86,67 @@ class SymptomInfo extends React.Component {
    *
    * @param {Event} event
    */
-  handleInfoTextMouseLeave() {
-    this.mouseOverInfoText = false
-
-    this.setState({
-      showInfoText: false,
-    })
+  function handleInfoTextMouseLeave() {
+    mouseOverInfoText = false
+    setShowInfoText(false)
   }
 
   /**
    * Handles any window resizes.
    */
-  handleWindowResize() {
-    this.throttledCalculateLeftShift()
+  function handleWindowResize() {
+    throttledCalculateLeftShift()
   }
 
-  render() {
-    const { infoText } = this.props
-    const {
-      arrowLeft,
-      containerLeft,
-      showInfoText,
-    } = this.state
+  useEffect(
+    () => {
+      throttledCalculateLeftShift()
+      window.addEventListener('resize', handleWindowResize)
 
-    return (
+      return () => {
+        window.removeEventListener('resize', handleWindowResize)
+      }
+    },
+    [symptomInfoContainer.current]
+  )
+
+  useEffect(() => {
+    return () => {
+      if (showInfoTextTimeout) {
+        window.clearTimeout(showInfoTextTimeout)
+      }
+    }
+  })
+
+  return (
+    <div
+      ref={symptomInfoContainer}
+      className="symptom-info-container"
+      onMouseEnter={handleInfoTextMouseEnter}
+      onMouseLeave={handleInfoTextMouseLeave}
+    >
+      <span onClick={handleInfoTextClick}>
+        <InfoIcon />
+      </span>
+
       <div
-        ref={symptomInfoContainer => this.symptomInfoContainer = symptomInfoContainer}
-        className="symptom-info-container"
-        onMouseEnter={this.handleInfoTextMouseEnter}
-        onMouseLeave={this.handleInfoTextMouseLeave}
+        className={cn('symptom-info-content-container', {
+          hidden: !showInfoText,
+        })}
+        style={{
+          left: containerLeft,
+        }}
       >
-        <span onClick={this.handleInfoTextClick}>
-          <InfoIcon />
-        </span>
-
         <div
-          className={cn('symptom-info-content-container', {
-            hidden: !showInfoText,
-          })}
+          className="symptom-info-arrow-indicator"
           style={{
-            left: containerLeft,
+            left: arrowLeft,
           }}
-        >
-          <div
-            className="symptom-info-arrow-indicator"
-            style={{
-              left: arrowLeft,
-            }}
-          />
-          <div className="symptom-info-content">
-            {infoText}
-          </div>
-        </div>
+        />
+        <div className="symptom-info-content">{infoText}</div>
       </div>
-    )
-  }
+    </div>
+  )
 }
 
 SymptomInfo.propTypes = {
