@@ -1,15 +1,12 @@
-// We import all of babel-polyfill here but the babel-preset-env
-// "useBuiltIns" option takes care of only including what we need.
-import 'babel-polyfill'
-
 import React from 'react'
 import ReactDOM from 'react-dom'
-import { AppContainer } from 'react-hot-loader'
+import { AppContainer, setConfig } from 'react-hot-loader'
 
-import { configureStore } from './store'
 import Root from './Root'
 
-const store = configureStore()
+// Currently an issue with React Hooks and React Hot Loader. This seems to fix the issue.
+// See https://github.com/gaearon/react-hot-loader/issues/1088#issuecomment-434862175.
+setConfig({ pureSFC: true })
 
 /**
  * Renders our React root wrapped with a hot-reloading component (NOTE: That component is a no-op in prod).
@@ -17,9 +14,7 @@ const store = configureStore()
 const render = () => {
   ReactDOM.render(
     <AppContainer>
-      <Root
-        store={store}
-      />
+      <Root />
     </AppContainer>,
     document.getElementById('app')
   )
