@@ -1,19 +1,19 @@
-const path = require('path')
+const path = require("path");
 
 module.exports = {
   serverPort: 9003,
   paths: {
-    base: path.resolve(__dirname, '.'),
+    base: path.resolve(__dirname, "."),
 
-    babelCache: path.resolve(__dirname, '.babel-cache'),
+    babelCache: path.resolve(__dirname, ".babel-cache"),
 
-    dist: path.resolve(__dirname, 'dist'),
+    dist: path.resolve(__dirname, "dist"),
 
-    app: {
-      base: path.resolve(__dirname, 'app'),
-      favicon: path.resolve(__dirname, 'app/assets/images/favicon.ico'),
-      mainJs: path.resolve(__dirname, 'app/main.jsx'),
-      html: path.resolve(__dirname, 'app/views/index.html'),
-    },
-  },
-}
+    src: {
+      base: path.resolve(__dirname, "app"),
+      favicon: path.resolve(__dirname, "app/assets/images/favicon.ico"),
+      mainJs: path.resolve(__dirname, "app/main.jsx"),
+      html: path.resolve(__dirname, "app/views/index.html")
+    }
+  }
+};
