@@ -3,7 +3,7 @@ const autoprefixer = require('autoprefixer')
 module.exports = {
   plugins: [
     autoprefixer({
-      browsers: ['> 5%'],
+      browsers: ['> 1%'],
       cascade: false,
     }),
   ],
