@@ -1,98 +1,96 @@
-const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin
-const DotenvPlugin = require('webpack-dotenv-plugin')
-const HtmlWebpackPlugin = require('html-webpack-plugin')
-const webpack = require('webpack')
+const DotenvPlugin = require("webpack-dotenv-plugin");
+const HtmlWebpackPlugin = require("html-webpack-plugin");
+const webpack = require("webpack");
 
-const buildConfig = require('./buildConfig')
+const buildConfig = require("./buildConfig");
 
 module.exports = {
   context: __dirname,
-  entry: [
-    'react-hot-loader/patch',
-    `webpack-dev-server/client?http://localhost:${buildConfig.serverPort}`,
-    'webpack/hot/only-dev-server',
-    buildConfig.paths.app.mainJs,
-  ],
-  output: {
-    chunkFilename: 'js/[name].js',
-    filename: 'js/[name].js',
-    path: buildConfig.paths.dist,
-    publicPath: '/',
+  devServer: {
+    clientLogLevel: "error", // The default value for this outputs too much in DevTools.
+    contentBase: buildConfig.paths.src.base,
+    historyApiFallback: {
+      disableDotRule: true
+    },
+    host: "0.0.0.0",
+    hot: true,
+    port: buildConfig.serverPort
   },
+  entry: [
+    "react-hot-loader/patch",
+    `webpack-dev-server/client?http://localhost:${buildConfig.serverPort}`,
+    "webpack/hot/only-dev-server",
+    buildConfig.paths.src.mainJs
+  ],
+  mode: "development",
   module: {
     rules: [
       {
         test: /\.jsx?$/,
-        include: [
-          buildConfig.paths.app.base,
-        ],
-        loader: 'babel-loader',
+        include: [buildConfig.paths.src.base],
+        loader: "babel-loader",
         options: {
-          cacheDirectory: buildConfig.paths.babelCache,
-        },
+          cacheDirectory: buildConfig.paths.babelCache
+        }
       },
       {
         test: /\.scss$/,
-        include: [
-          buildConfig.paths.app.base,
-        ],
-        use: [
-          'style-loader',
-          'css-loader',
-          'sass-loader',
-        ],
+        include: [buildConfig.paths.src.base],
+        use: ["style-loader", "css-loader", "sass-loader"]
       },
-    ],
+      {
+        test: /\.(woff(2)?|ttf|eot|svg)(\?v=\d+\.\d+\.\d+)?$/,
+        loader: "file-loader",
+        options: {
+          name: "[name].[hash].[ext]"
+        }
+      }
+    ]
+  },
+  optimization: {
+    splitChunks: {
+      cacheGroups: {
+        commons: {
+          chunks: "all",
+          name: "vendor",
+          test: /[\\/]node_modules[\\/]/
+        }
+      }
+    },
+    runtimeChunk: {
+      name: "manifest"
+    }
+  },
+  output: {
+    chunkFilename: "js/[name].js",
+    filename: "js/[name].js",
+    path: buildConfig.paths.dist,
+    publicPath: "/"
   },
   plugins: [
-    new BundleAnalyzerPlugin({
-      openAnalyzer: false,
-    }),
-
+    // We don't really use this as different env files for different people. Just a place to keep common
+    // env variables for the project as a whole.
     new DotenvPlugin({
-      // We don't really use this as different env files for different people. Just a place to keep common
-      // env variables for the project as a whole.
       allowEmptyValues: true,
-      sample: '.env',
-      path: '.env',
+      path: ".env",
+      sample: ".env"
     }),
     new HtmlWebpackPlugin({
-      favicon: buildConfig.paths.app.favicon,
-      template: buildConfig.paths.app.html,
+      favicon: buildConfig.paths.src.favicon,
+      // "inject: true" places all JavaScript resources at the bottom of the body element.
+      inject: true,
+      template: buildConfig.paths.src.html
     }),
     new webpack.HotModuleReplacementPlugin(),
 
     // Allows the HMR plugin to output more legible names.
     new webpack.NamedModulesPlugin(),
 
-    // This makes our vendor bundle from node_modules modules.
-    new webpack.optimize.CommonsChunkPlugin({
-      name: 'vendor',
-      minChunks(module) {
-        return module.context && module.context.indexOf('node_modules') !== -1
-      },
-    }),
-
-    // This ensures that our vendor bundle name doesn't change between builds (unless the vendor contents change)
-    // by extracting out the webpack bootstrap code into its own file.
-    new webpack.optimize.CommonsChunkPlugin({
-      name: 'manifest',
-      minChunks: Infinity,
-    }),
+    // Ignore locales from moment.
+    new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/)
   ],
   resolve: {
-    modules: [
-      'node_modules',
-      buildConfig.paths.base,
-    ],
-    extensions: [
-      '.js',
-      '.jsx',
-    ],
-  },
-  devServer: {
-    historyApiFallback: true,
-    hot: true,
-    port: buildConfig.serverPort,
-  },
-}
+    extensions: [".js", ".jsx"],
+    modules: ["node_modules", buildConfig.paths.base]
+  }
+};
