@@ -1,109 +1,97 @@
-const CopyWebpackPlugin = require('copy-webpack-plugin')
-const ExtractTextPlugin = require('extract-text-webpack-plugin')
-const HtmlWebpackPlugin = require('html-webpack-plugin')
-const webpack = require('webpack')
+const CopyWebpackPlugin = require("copy-webpack-plugin");
+const HtmlWebpackPlugin = require("html-webpack-plugin");
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+const webpack = require("webpack");
 
-const buildConfig = require('./buildConfig')
+const buildConfig = require("./buildConfig");
 
 module.exports = {
   bail: true,
   context: __dirname,
   entry: {
-    main: buildConfig.paths.app.mainJs,
+    main: buildConfig.paths.src.mainJs
   },
-  output: {
-    chunkFilename: 'js/[name].[chunkhash].js',
-    filename: 'js/[name].[chunkhash].js',
-    path: buildConfig.paths.dist,
-    publicPath: '/',
-  },
+  mode: "production",
   module: {
     rules: [
       {
         test: /\.jsx?$/,
-        include: [
-          buildConfig.paths.app.base,
-        ],
-        loader: 'babel-loader',
+        include: [buildConfig.paths.src.base],
+        loader: "babel-loader"
       },
       {
         test: /\.scss$/,
-        include: [
-          buildConfig.paths.app.base,
-        ],
-        loader: ExtractTextPlugin.extract({
-          fallback: 'style-loader',
-          use: [
-            'css-loader?importLoaders=1',
-            'postcss-loader',
-            'sass-loader',
-          ],
-        }),
+        include: [buildConfig.paths.src.base],
+        use: [
+          MiniCssExtractPlugin.loader,
+          "css-loader?importLoaders=1",
+          "postcss-loader",
+          "sass-loader"
+        ]
       },
-    ],
+      {
+        test: /\.(woff(2)?|ttf|eot|svg)(\?v=\d+\.\d+\.\d+)?$/,
+        loader: "file-loader",
+        options: {
+          name: "[name].[hash].[ext]"
+        }
+      }
+    ]
+  },
+  optimization: {
+    splitChunks: {
+      cacheGroups: {
+        commons: {
+          chunks: "all",
+          name: "vendor",
+          test: /[\\/]node_modules[\\/]/
+        }
+      }
+    },
+    runtimeChunk: {
+      name: "manifest"
+    }
+  },
+  output: {
+    chunkFilename: "js/[name].[chunkhash].js",
+    filename: "js/[name].[chunkhash].js",
+    path: buildConfig.paths.dist,
+    publicPath: "/"
   },
   plugins: [
     // This is a shorthand plugin for the DefinePlugin.
     new webpack.EnvironmentPlugin([
-      'APP_ENV',
-      'GA_PROPERTY',
-      'INSPECTLET_APP_ID',
-      'NODE_ENV',
+      "APP_ENV",
+      "GA_PROPERTY",
+      "INSPECTLET_APP_ID",
+      "NODE_ENV"
     ]),
-
     new HtmlWebpackPlugin({
-      favicon: buildConfig.paths.app.favicon,
-      template: buildConfig.paths.app.html,
+      favicon: buildConfig.paths.src.favicon,
+      // "inject: true" places all JavaScript resources at the bottom of the body element.
+      inject: true,
+      template: buildConfig.paths.src.html
     }),
+
+    // Ignore locales from moment.
+    new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/),
 
     // Need this to preserve the IDs of Webpack modules between builds. Otherwise having new imports in the main bundle
     // will cache-bust the vendor bundle.
     new webpack.HashedModuleIdsPlugin(),
 
-    // Enable scope-hoisting.
-    new webpack.optimize.ModuleConcatenationPlugin(),
-
-    // This makes our vendor bundle from node_modules modules.
-    new webpack.optimize.CommonsChunkPlugin({
-      minChunks: module => module.context && module.context.indexOf('node_modules') !== -1,
-      name: 'vendor',
+    new MiniCssExtractPlugin({
+      filename: "css/[name].[contenthash].css"
     }),
 
-    // This ensures that our vendor bundle name doesn't change between builds (unless the vendor contents change)
-    // by extracting out the webpack bootstrap code into its own file.
-    new webpack.optimize.CommonsChunkPlugin({
-      name: 'manifest',
-      minChunks: Infinity,
-    }),
-
-    new webpack.optimize.UglifyJsPlugin({
-      compress: {
-        screw_ie8: true, // eslint-disable-line camelcase
-        warnings: false,
-      },
-      mangle: {
-        screw_ie8: true, // eslint-disable-line camelcase
-      },
-      output: {
-        comments: false,
-        screw_ie8: true, // eslint-disable-line camelcase
-      },
-    }),
-
-    new ExtractTextPlugin('css/[name].[contenthash].css'),
-
-    new CopyWebpackPlugin([{
-      from: './_redirects',
-    }]),
+    new CopyWebpackPlugin([
+      {
+        from: "./_redirects"
+      }
+    ])
   ],
   resolve: {
-    modules: [
-      'node_modules',
-      buildConfig.paths.base,
-    ],
-    extensions: [
-      '.js',
-      '.jsx',
-    ],
-  },
-}
+    extensions: [".js", ".jsx"],
+    modules: ["node_modules", buildConfig.paths.base]
+  }
+};
